features/chatbot/in house (ayush and shravani) 3/utils/tts.py
from google.cloud import texttospeech
from io import BytesIO

def synthesize_speech(text: str, language: str = "en-US") -> BytesIO:
    client = texttospeech.TextToSpeechClient()

    synthesis_input = texttospeech.SynthesisInput(text=text)
    voice = texttospeech.VoiceSelectionParams(
        language_code=language,
        ssml_gender=texttospeech.SsmlVoiceGender.NEUTRAL,
    )
    audio_config = texttospeech.AudioConfig(
        audio_encoding=texttospeech.AudioEncoding.MP3
    )

    response = client.synthesize_speech(
        input=synthesis_input, voice=voice, audio_config=audio_config
    )

    audio_stream = BytesIO(response.audio_content)
    audio_stream.seek(0)
    return audio_stream

