features/chatbot/in house (ayush and shravani) 3/requirements.txt
python-dotenv
assemblyai
pyttsx3
google-cloud-texttospeech
google-cloud-speech
pydub
openai==0.28
fastapi
uvicorn
numpy
pandas
pickle-mixin
python-dotenv
